Goes with Chip implementation of party view with a card backgorund
